# apiEscola/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "AppEscola API",
        "description": "API REST do sistema multi-tenant AppEscola — gestão de escolas e cursinhos.",
        "contact": {
            "email": "contato@appescola.com.br"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:4000",
            "description": "Ambiente local"
        }
    ],
    "paths": {
...
        "/api/course-bundles": {
            "get": {
                "tags": [
                    "CourseBundles"
                ],
                "summary": "Listar pacotes de cursos",
                "operationId": "14606a0195a377428e469b495580d787",
                "responses": {
                    "200": {
                        "description": "Lista de pacotes"
                    },
                    "401": {
                        "description": "Não autenticado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "CourseBundles"
                ],
                "summary": "Criar pacote de cursos",
                "operationId": "7191619980c2cc9653e8798a38d5e53e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "billing_cycle",
                                    "price",
                                    "course_ids"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "CPM Completo"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "billing_cycle": {
                                        "type": "string",
                                        "example": "monthly",
                                        "enum": [
                                            "monthly",
                                            "bimonthly",
                                            "quadrimestral",
                                            "semiannual",
                                            "annual"
                                        ]
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 170
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "active"
                                    },
                                    "course_ids": {
                                        "description": "IDs dos cursos que compõem o pacote",
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Pacote criado"
                    },
                    "422": {
                        "description": "Dados inválidos"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/course-bundles/{id}": {
            "get": {
                "tags": [
                    "CourseBundles"
                ],
                "summary": "Exibir pacote",
                "operationId": "87d7e12365fc7a68025ee000c3f2c9a3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dados do pacote"
                    },
                    "404": {
                        "description": "Não encontrado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "CourseBundles"
                ],
                "summary": "Atualizar pacote",
                "operationId": "f67c9a54751ad55119734d20b3b9bd64",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "description": "Campos a atualizar"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Pacote atualizado"
                    },
                    "422": {
                        "description": "Dados inválidos"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "CourseBundles"
                ],
                "summary": "Remover pacote",
                "operationId": "4e4796544d12e21e387537ba0aaf2cc8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Removido com sucesso"
                    },
                    "404": {
                        "description": "Não encontrado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Informe o token retornado pelo endpoint /api/login",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}